pyproject.toml
[project]
name = "mcp-implement"
version = "0.1.0"
description = "autoMD"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair<5",
    "anthropic>=0.86.0",
    "fastmcp>=3.2.0",
    "google-generativeai>=0.8.6",
    "mcp[cli]>=1.26.0",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "plotly-express>=0.4.1",
    "px>=0.2.2",
    "pytest>=7.4.0",
    "setuptools>=82.0.1",
    "streamlit==1.35.0",
    "streamlit-plotly-events>=0.0.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_implement*"]
